users: correct access patterns

+++ users/users.go
@@ -6,6 +6,7 @@ import (
 
 	"golang.org/x/crypto/bcrypt"
 
+	"github.com/appleboy/gin-jwt"
 	"github.com/gin-gonic/gin"
 	"github.com/jinzhu/gorm"
 
@@ -13,6 +14,7 @@ import (
 	"github.com/hellerve/artifex/model"
 )
 
+// User is the user model
 type User struct {
 	model.Base
 	Email     string   `json:"email" binding:"required"`
@@ -28,6 +30,7 @@ type User struct {
 	Social    *Social  `json:"social"`
 }
 
+// Address is a user’s address
 type Address struct {
 	model.Base
 	Line1 string `json:"line1"`
@@ -36,6 +39,7 @@ type Address struct {
 	State string `json:"state"`
 }
 
+// Social is a user’s social media accounts
 type Social struct {
 	model.Base
 	Github  string `json:"github"`
@@ -43,6 +47,7 @@ type Social struct {
 	Website string `json:"website"`
 }
 
+// Initialize initializes the URLs for users
 func Initialize(db *gorm.DB, router *gin.Engine, auth func() gin.HandlerFunc) {
 	router.POST("/users", endpoints.Endpoint(db, CreateUser))
 	g := router.Group("/users")
@@ -57,12 +62,14 @@ func Initialize(db *gorm.DB, router *gin.Engine, auth func() gin.HandlerFunc) {
 	db.AutoMigrate(&User{}, &Address{}, &Social{})
 }
 
+// GetUsers gets all users
 func GetUsers(c *gin.Context, db *gorm.DB) {
 	var users []User
 	db.Find(&users)
 	c.JSON(http.StatusOK, users)
 }
 
+// GetUser gets a specifc user
 func GetUser(c *gin.Context, db *gorm.DB) {
 	id, err := strconv.Atoi(c.Param("id"))
 
@@ -82,6 +89,7 @@ func GetUser(c *gin.Context, db *gorm.DB) {
 	c.JSON(http.StatusOK, user)
 }
 
+// CreateUser creates a new user
 func CreateUser(c *gin.Context, db *gorm.DB) {
 	var user User
 	if err := c.ShouldBindJSON(&user); err != nil {
@@ -113,6 +121,7 @@ func CreateUser(c *gin.Context, db *gorm.DB) {
 	c.JSON(http.StatusOK, user)
 }
 
+// DeleteUser deletes a user (must be the user themselves)
 func DeleteUser(c *gin.Context, db *gorm.DB) {
 	id, err := strconv.Atoi(c.Param("id"))
 
@@ -129,11 +138,21 @@ func DeleteUser(c *gin.Context, db *gorm.DB) {
 		return
 	}
 
+	claims := jwt.ExtractClaims(c)
+	var me User
+	db.Where("username = ?", claims["id"]).First(&me)
+
+	if user.ID != me.ID {
+		c.String(http.StatusForbidden, "Cannot alter foreign user")
+		return
+	}
+
 	db.Delete(&user)
 
 	c.String(http.StatusOK, "")
 }
 
+// UpdateUser updates a user (must be the user themselves)
 func UpdateUser(c *gin.Context, db *gorm.DB) {
 	id, err := strconv.Atoi(c.Param("id"))
 
@@ -158,6 +177,15 @@ func UpdateUser(c *gin.Context, db *gorm.DB) {
 
 	if (user.Staff && !other.Staff) || (user.Admin && !other.Admin) {
 		c.String(http.StatusForbidden, "Cannot make user admin")
+		return
+	}
+
+	claims := jwt.ExtractClaims(c)
+	var me User
+	db.Where("username = ?", claims["id"]).First(&me)
+
+	if user.ID != me.ID {
+		c.String(http.StatusForbidden, "Cannot alter foreign user")
 		return
 	}
